requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
arch==8.0.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.3.0
attrs==26.1.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
choreographer==1.2.1
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
dateparser==1.4.0
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
docxcompose==2.1.0
docxtpl==0.20.2
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.135.2
fastjsonschema==2.21.2
fonttools==4.61.1
fqdn==1.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
ipykernel==7.2.0
ipython==8.39.0
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.14.0
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kaleido==1.2.0
kiwisolver==1.4.9
lark==1.3.1
logistro==2.0.1
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.0
matplotlib-inline==0.2.1
mistune==3.2.0
narwhals==2.18.1
nbclient==0.10.4
nbconvert==7.17.1
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.5.5
notebook_shim==0.2.4
numpy==1.26.4
openpyxl==3.1.5
orjson==3.11.5
overrides==7.7.0
packaging==26.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.6
patsy==1.0.2
PeakUtils==1.3.5
pillow==12.1.1
platformdirs==4.9.6
plotly==6.6.0
pluggy==1.6.0
prometheus_client==0.25.0
prompt_toolkit==3.0.52
psutil==7.2.2
psycopg2==2.9.11
pure_eval==0.2.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pyparsing==3.3.2
pytest==9.0.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-json-logger==4.1.0
python-multipart==0.0.22
pytz==2025.2
pywinpty==3.0.3
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
scikit-base==0.13.0
scikit-learn==1.7.2
scipy==1.12.0
seaborn==0.13.2
Send2Trash==2.1.0
simplejson==3.20.2
six==1.17.0
sktime==0.40.1
soupsieve==2.8.3
stack-data==0.6.3
starlette==1.0.0
statsmodels==0.14.6
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.3.0
tornado==6.5.5
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.42.0
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.15
